storage/requirements.txt
connexion[flask,uvicorn,swagger-ui]
pykafka
sqlalchemy
mysqlclient
setuptools
pymysql
kafka-python
